FEBioApp: Modified CFloatSlider so that it displays a slider and an (read only) QLineEdit.
git-svn-id: https://gforge.sci.utah.edu/svn/MRLProjects/FEBioApp@10804 b505bdfe-9d5c-0410-8600-697d5fc8e487

+++ FEBioApp/UIBuilder.cpp
@@ -683,7 +683,7 @@ void UIBuilder::parseInput(XMLTag& tag, QBoxLayout* playout)
 			slider->setFloatRange(rng[0], rng[1], rng[2]);
 			pw = slider;
 			pi->SetWidget(slider);
-			QObject::connect(slider, SIGNAL(valueChanged(int)), m_dlg, SLOT(paramChanged()));
+			QObject::connect(slider, SIGNAL(valueChanged(double)), m_dlg, SLOT(paramChanged()));
 		}
 		else
 		{
